Make game scalable
This change adds a scale function on the main `game` object and uses the
preset values (1600 x 1200) which were chosen as dimensions for the game
to find the relevant values based on the actual width and height assigned
to the game. This will allow the user to define any dimensions on the
parent div and the game will adapt accordingly.

Close #16